+++ CHANGELOG.md
@@ -6,3 +6,8 @@
 - ensure all item ids are unique
 - ability to choose the colors for everything
 
+## 0.1.0
+- nested items (N levels)
+- each item now receives a ListTile component to facilite color customizations
+- global (apply to all items) color configuration for states like hover, focus, selected.
+